+++ Game/Game.vcxproj
@@ -21,6 +21,9 @@
   <ItemGroup>
     <ClCompile Include="GameState.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="GameState.h" />
+  </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>17.0</VCProjectVersion>
     <Keyword>Win32Proj</Keyword>
